Code clean up
Fix Inconsistent naming, change all types to CLR, reorder modifiers, use explicit modifiers, add readionly modifiers. Add abreviations to team-shared resharper abreviation list.

+++ PacketDotNet/OSPF/TOSMetric.cs
@@ -1,3 +1,4 @@
+using System;
 using PacketDotNet.Utils.Conversion;
 
 namespace PacketDotNet.OSPF
@@ -9,28 +10,28 @@ namespace PacketDotNet.OSPF
     public struct TOSMetric
     {
         ///<summary>The number of bytes a TOS metric occupy</summary>
-        public static readonly int TOSMetricLength = 4;
+        public static readonly Int32 TOSMetricLength = 4;
 
         /// <summary>
         /// IP Type of Service that this metric refers to.
         /// </summary>
-        public byte TOS;
+        public Byte TOS;
 
         /// <summary>
         /// TOS-specific metric information.
         /// </summary>
-        public uint Metric;
+        public UInt32 Metric;
 
         /// <summary>
         /// Gets the bytes that make up this packet.
         /// </summary>
         /// <value>Packet bytes</value>
-        public byte[] Bytes
+        public Byte[] Bytes
         {
             get
             {
-                byte[] b = new byte[TOSMetricLength];
-                EndianBitConverter.Big.CopyBytes((uint) this.Metric, b, 0);
+                Byte[] b = new Byte[TOSMetricLength];
+                EndianBitConverter.Big.CopyBytes(this.Metric, b, 0);
                 b[0] = this.TOS;
                 return b;
             }
